requirements.txt
alabaster==0.7.9
Babel==2.3.4
docutils==0.12
imagesize==0.7.1
Jinja2==2.8
MarkupSafe==0.23
numpy==1.11.2
pathlib2==2.1.0
pep8==1.7.0
pickleshare==0.7.4
pkg-resources==0.0.0
pygame==1.9.2b8
Pygments==2.1.3
pyserial==3.2.1
pytz==2016.7
QtPy==1.1.2
scipy==0.18.1
six==1.10.0
snowballstemmer==1.2.1
Sphinx==1.4.8
